refactor: Seperate two tabs into its own file
