+++ lib/presentation/user/home/roles_tab.dart
@@ -1,0 +1,223 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:mobile_ai_erp/domain/entity/user/role.dart';
+import 'package:mobile_ai_erp/core/services/tenant_service.dart';
+import 'package:mobile_ai_erp/presentation/user/store/role_store.dart';
+
+class RolesTab extends StatefulWidget {
+  final RoleStore roleStore;
+
+  const RolesTab({
+    super.key,
+    required this.roleStore,
+  });
+
+  @override
+  State<RolesTab> createState() => _RolesTabState();
+}
+
+class _RolesTabState extends State<RolesTab> {
+  @override
+  void initState() {
+    super.initState();
+    widget.roleStore.loadRoles();
+  }
+
+  void _showCreateRoleDialog(BuildContext context) {
+    final nameController = TextEditingController();
+    final descController = TextEditingController();
+
+    showDialog(
+      context: context,
+      builder: (_) {
+        return AlertDialog(
+          title: const Text('Create Role'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextField(
+                controller: nameController,
+                decoration: const InputDecoration(labelText: 'Role name'),
+              ),
+              TextField(
+                controller: descController,
+                decoration: const InputDecoration(labelText: 'Description'),
+              ),
+            ],
+          ),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(context),
+              child: const Text('Cancel'),
+            ),
+            StatefulBuilder(
+              builder: (context, setState) {
+                return Observer(
+                  builder: (_) {
+                    return ElevatedButton(
+                      onPressed: widget.roleStore.loading ? null : () async {
+                        final tenantService = TenantService();
+                        final tenantId = await tenantService.getCurrentTenantId();
+                        
+                        if (tenantId == null) {
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            const SnackBar(content: Text('Unable to get tenant ID')),
+                          );
+                          return;
+                        }
+
+                        if (nameController.text.trim().isEmpty) {
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            const SnackBar(content: Text('Please enter a role name')),
+                          );
+                          return;
+                        }
+
+                        final newRole = Role(
+                          id: DateTime.now().millisecondsSinceEpoch.toString(),
+                          tenantId: tenantId,
+                          name: nameController.text.trim(),
+                          description: descController.text.trim(),
+                        );
+
+                        await widget.roleStore.createRole(newRole);
+                        
+                        if (widget.roleStore.error == null) {
+                          Navigator.pop(context);
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            const SnackBar(content: Text('Role created successfully')),
+                          );
+                        } else {
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(
+                              content: Text(widget.roleStore.error!),
+                              backgroundColor: Colors.red,
+                            ),
+                          );
+                        }
+                      },
+                      child: widget.roleStore.loading 
+                        ? const SizedBox(
+                            width: 16,
+                            height: 16,
+                            child: CircularProgressIndicator(strokeWidth: 2),
+                          )
+                        : const Text('Create'),
+                    );
+                  },
+                );
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Observer(
+      builder: (_) {
+        if (widget.roleStore.loading) {
+          return const Center(child: CircularProgressIndicator());
+        }
+
+        if (widget.roleStore.error != null) {
+          return Center(
+            child: Padding(
+              padding: const EdgeInsets.all(16),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Icon(
+                    Icons.error_outline,
+                    size: 64,
+                    color: Theme.of(context).colorScheme.error,
+                  ),
+                  const SizedBox(height: 16),
+                  Text(
+                    'Something went wrong',
+                    style: Theme.of(context).textTheme.headlineSmall,
+                    textAlign: TextAlign.center,
+                  ),
+                  const SizedBox(height: 8),
+                  Text(
+                    widget.roleStore.error!,
+                    style: Theme.of(context).textTheme.bodyMedium,
+                    textAlign: TextAlign.center,
+                  ),
+                  const SizedBox(height: 24),
+                  ElevatedButton.icon(
+                    onPressed: () => widget.roleStore.loadRoles(),
+                    icon: const Icon(Icons.refresh),
+                    label: const Text('Retry'),
+                  ),
+                ],
+              ),
+            ),
+          );
+        }
+
+        return Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Align(
+                alignment: Alignment.centerRight,
+                child: ElevatedButton.icon(
+                  onPressed: () => _showCreateRoleDialog(context),
+                  icon: const Icon(Icons.add),
+                  label: const Text('Create Role'),
+                ),
+              ),
+            ),
+            Expanded(
+              child: ListView.builder(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                itemCount: widget.roleStore.roleList.length,
+                itemBuilder: (_, index) {
+                  final role = widget.roleStore.roleList[index];
+
+                  return Card(
+                    margin: const EdgeInsets.only(bottom: 16),
+                    child: Padding(
+                      padding: const EdgeInsets.all(16),
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              Text(
+                                role.name,
+                                style: Theme.of(context).textTheme.titleMedium,
+                              ),
+                            ],
+                          ),
+                          const SizedBox(height: 8),
+                          Text(
+                            role.description ?? 'No description',
+                            style: const TextStyle(color: Colors.grey),
+                          ),
+                          const SizedBox(height: 12),
+                          Row(
+                            children: [
+                              IconButton(
+                                onPressed: () => widget.roleStore.deleteRole(role.id),
+                                icon: const Icon(Icons.delete),
+                              ),
+                            ],
+                          )
+                        ],
+                      ),
+                    ),
+                  );
+                },
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
