+++ lib/presentation/user/home/user_home.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:mobile_ai_erp/domain/entity/user/role.dart';
 import 'package:mobile_ai_erp/domain/entity/user/user.dart';
 import 'package:mobile_ai_erp/presentation/user/store/role_store.dart';
 import 'package:mobile_ai_erp/presentation/user/store/user_store.dart';
 import 'package:mobile_ai_erp/domain/entity/user/user_status.dart';
-import 'package:mobile_ai_erp/core/services/tenant_service.dart';
+import 'package:mobile_ai_erp/presentation/user/home/users_tab.dart';
+import 'package:mobile_ai_erp/presentation/user/home/roles_tab.dart';
 
 class UserManagementScreen extends StatefulWidget {
   final UserStore userStore;
@@ -40,13 +40,6 @@ class _UserManagementScreenState extends State<UserManagementScreen>
     _tabController = TabController(length: 2, vsync: this);
   }
 
-  Role? _findRoleById(String id) {
-    return widget.roleStore.roleList
-        .where((r) => r.id == id)
-        .cast<Role?>()
-        .firstOrNull;
-  }
-
   @override
   void dispose() {
     _tabController.dispose();
@@ -73,83 +66,14 @@ class _UserManagementScreenState extends State<UserManagementScreen>
       body: TabBarView(
         controller: _tabController,
         children: [
-          _buildUserTab(),
-          _buildRoleTab(),
+          UsersTab(
+            userStore: widget.userStore,
+            roleStore: widget.roleStore,
+            onGrantRole: () => _showGrantRoleDialog(context, widget.userStore.userList.first),
+          ),
+          RolesTab(roleStore: widget.roleStore),
         ],
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          if (_tabController.index == 0) {
-            _showCreateUserDialog(context);
-          } else {
-            _showCreateRoleDialog(context);
-          }
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-
-  Widget _buildUserTab() {
-    return Observer(
-      builder: (_) {
-        if (widget.userStore.loading) {
-          return const Center(child: CircularProgressIndicator());
-        }
-
-        return ListView.builder(
-          padding: const EdgeInsets.all(16),
-          itemCount: widget.userStore.userList.length,
-          itemBuilder: (_, index) {
-            final user = widget.userStore.userList[index];
-            final role = _findRoleById(user.roleId);
-
-            return Card(
-              margin: const EdgeInsets.only(bottom: 16),
-              child: Padding(
-                padding: const EdgeInsets.all(16),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Text(
-                          user.name,
-                          style: Theme.of(context).textTheme.titleMedium,
-                        ),
-                        _StatusBox(
-                          isActive: user.status == UserStatus.active,
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 8),
-                    Text(user.email),
-                    const SizedBox(height: 12),
-                    Chip(
-                      label: Text(role?.name ?? 'No role'),
-                    ),
-                    const SizedBox(height: 12),
-                    Row(
-                      children: [
-                        ElevatedButton(
-                          onPressed: () => _showGrantRoleDialog(context, user),
-                          child: const Text('Grant Role'),
-                        ),
-                        const SizedBox(width: 8),
-                        IconButton(
-                          onPressed: () => widget.userStore.deleteUser(user.id),
-                          icon: const Icon(Icons.delete),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            );
-          },
-        );
-      },
     );
   }
 
@@ -204,291 +128,4 @@ class _UserManagementScreenState extends State<UserManagementScreen>
     );
   }
 
-  void _showCreateUserDialog(BuildContext context) {
-    final nameController = TextEditingController();
-    final emailController = TextEditingController();
-    final phoneController = TextEditingController();
-
-    String? selectedRoleId;
-
-    showDialog(
-      context: context,
-      builder: (_) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            return AlertDialog(
-              title: const Text('Create User'),
-              content: SingleChildScrollView(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    TextField(
-                      controller: nameController,
-                      decoration: const InputDecoration(labelText: 'Name'),
-                    ),
-                    TextField(
-                      controller: emailController,
-                      decoration: const InputDecoration(labelText: 'Email'),
-                    ),
-                    TextField(
-                      controller: phoneController,
-                      decoration: const InputDecoration(labelText: 'Phone'),
-                    ),
-                    const SizedBox(height: 16),
-                    Observer(
-                      builder: (_) {
-                        return DropdownButtonFormField<String>(
-                          hint: const Text('Select Role'),
-                          value: selectedRoleId,
-                          items: widget.roleStore.roleList.map((role) {
-                            return DropdownMenuItem(
-                              value: role.id,
-                              child: Text(role.name),
-                            );
-                          }).toList(),
-                          onChanged: (value) {
-                            setState(() {
-                              selectedRoleId = value;
-                            });
-                          },
-                        );
-                      },
-                    ),
-                  ],
-                ),
-              ),
-              actions: [
-                TextButton(
-                  onPressed: () => Navigator.pop(context),
-                  child: const Text('Cancel'),
-                ),
-                ElevatedButton(
-                  onPressed: () {
-                    if (selectedRoleId == null) return;
-
-                    final newUser = User(
-                      id: DateTime.now().millisecondsSinceEpoch,
-                      name: nameController.text,
-                      email: emailController.text,
-                      phone: phoneController.text,
-                      status: UserStatus.active,
-                      roleId: selectedRoleId!,
-                    );
-
-                    widget.userStore.createUser(newUser);
-                    Navigator.pop(context);
-                  },
-                  child: const Text('Create'),
-                ),
-              ],
-            );
-          },
-        );
-      },
-    );
-  }
-
-  Widget _buildRoleTab() {
-    return Observer(
-      builder: (_) {
-        if (widget.roleStore.loading) {
-          return const Center(child: CircularProgressIndicator());
-        }
-
-        if (widget.roleStore.error != null) {
-          return Center(
-            child: Padding(
-              padding: const EdgeInsets.all(16),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(
-                    Icons.error_outline,
-                    size: 64,
-                    color: Theme.of(context).colorScheme.error,
-                  ),
-                  const SizedBox(height: 16),
-                  Text(
-                    'Something went wrong',
-                    style: Theme.of(context).textTheme.headlineSmall,
-                    textAlign: TextAlign.center,
-                  ),
-                  const SizedBox(height: 8),
-                  Text(
-                    widget.roleStore.error!,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                    textAlign: TextAlign.center,
-                  ),
-                  const SizedBox(height: 24),
-                  ElevatedButton.icon(
-                    onPressed: () => widget.roleStore.loadRoles(),
-                    icon: const Icon(Icons.refresh),
-                    label: const Text('Retry'),
-                  ),
-                ],
-              ),
-            ),
-          );
-        }
-
-        return ListView.builder(
-          padding: const EdgeInsets.all(16),
-          itemCount: widget.roleStore.roleList.length,
-          itemBuilder: (_, index) {
-            final role = widget.roleStore.roleList[index];
-
-            return Card(
-              margin: const EdgeInsets.only(bottom: 16),
-              child: Padding(
-                padding: const EdgeInsets.all(16),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        Text(
-                          role.name,
-                          style: Theme.of(context).textTheme.titleMedium,
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      role.description ?? 'No description',
-                      style: const TextStyle(color: Colors.grey),
-                    ),
-                    const SizedBox(height: 12),
-                    Row(
-                      children: [
-                        IconButton(
-                          onPressed: () => widget.roleStore.deleteRole(role.id),
-                          icon: const Icon(Icons.delete),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-
-  void _showCreateRoleDialog(BuildContext context) {
-    final nameController = TextEditingController();
-    final descController = TextEditingController();
-
-    showDialog(
-      context: context,
-      builder: (_) {
-        return AlertDialog(
-          title: const Text('Create Role'),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextField(
-                controller: nameController,
-                decoration: const InputDecoration(labelText: 'Role name'),
-              ),
-              TextField(
-                controller: descController,
-                decoration: const InputDecoration(labelText: 'Description'),
-              ),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(context),
-              child: const Text('Cancel'),
-            ),
-            StatefulBuilder(
-                builder: (context, setState) {
-                  return Observer(
-                    builder: (_) {
-                      return ElevatedButton(
-                        onPressed: widget.roleStore.loading ? null : () async {
-                          final tenantService = TenantService();
-                          final tenantId = await tenantService.getCurrentTenantId();
-                          
-                          if (tenantId == null) {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(content: Text('Unable to get tenant ID')),
-                            );
-                            return;
-                          }
-
-                          if (nameController.text.trim().isEmpty) {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(content: Text('Please enter a role name')),
-                            );
-                            return;
-                          }
-
-                          final newRole = Role(
-                            id: DateTime.now().millisecondsSinceEpoch.toString(),
-                            tenantId: tenantId,
-                            name: nameController.text.trim(),
-                            description: descController.text.trim(),
-                          );
-
-                          await widget.roleStore.createRole(newRole);
-                          
-                          if (widget.roleStore.error == null) {
-                            Navigator.pop(context);
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(content: Text('Role created successfully')),
-                            );
-                          } else {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              SnackBar(
-                                content: Text(widget.roleStore.error!),
-                                backgroundColor: Colors.red,
-                              ),
-                            );
-                          }
-                        },
-                        child: widget.roleStore.loading 
-                          ? const SizedBox(
-                              width: 16,
-                              height: 16,
-                              child: CircularProgressIndicator(strokeWidth: 2),
-                            )
-                          : const Text('Create'),
-                      );
-                    },
-                  );
-                },
-              ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-class _StatusBox extends StatelessWidget {
-  final bool isActive;
-
-  const _StatusBox({required this.isActive});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 4),
-      decoration: BoxDecoration(
-        color: isActive ? Colors.green.shade100 : Colors.red.shade100,
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Text(
-        isActive ? 'Active' : 'Inactive',
-        style: TextStyle(
-          color: isActive ? Colors.green : Colors.red,
-          fontWeight: FontWeight.w500,
-        ),
-      ),
-    );
-  }
-}
+    }
